rapids-cmake add rapids_cpm_<PackageName> for common packages
Fixes #32

The following packages are now are easier to user for RAPIDS projects
as rapids-cmake offers a pre-configured setup for each project.

  GTest
  NVBench
  RMM
  SpdLog
  Thrust

On top of providing a consistent version of these packages to
all RAPIDS projects, rapids-cmake now is able to deduce when these
projects should also be installed.

```cmake

rapids_cpm_gtest(BUILD_EXPORT_SET myproject)
rapdis_cpm_rmm(BUILD_EXPORT_SET myproject
               INSTALL_EXPORT_SET myproject)
```
Given the above snippet when RMM is built as a subcomponent of
`myproject` it will be installed as well. This is done since
RMM is part of the INSTALL export set, and therefore must
be available from an installed version of `myproject`.