rapids_export will always install cmake files to `lib`
Fixes #56
While using `GNUInstallDirs` is the technically correct way to
determine the correct install location for CMake based projects,
it has no concept of conda enviornments and incorrectly uses
the base distros layout instead.

In the very short term have `rapids-export` always export to `lib` as other
RAPIDS projects hardcode the install location to that

In the short term extend `rapids-cmake` to offer a function to compute
the library location which is conda aware. Update all of RAPIDS to
use that.

Medium term update CMake GNUInstallDirs to be conda aware so we
can just assume that works as expected.