Introduce rapids_cpm_libcudacxx
Provides an unified version of libcudacxx for all rapids-cmake based projects.

Since the libcudacxx is focused on building the project, we can't rely
on its install rules. Instead we manually produce install rules and a
libcudacxx-config.cmake to allow find_package(libcudacxx).

Like Thrust we also make sure the includes are namespaced and
always marked as user include so that consumers use this version
of libcudacxx instead of the one provided in the system cuda toolkit.

Fixes #97