Tracked package versions are now not required, but preferred.
Instead of generating `find_dependency(<pkg> <version>)` we
now generate:
  find_package(<pkg> <version>)
  find_dependency(<pkg>)

This allows us to search for the versioned type, but if it
doesn't exist fall back to an unversioned one. Since
find_dependency propagates the `REQUIRED` keyword we need
to use `find_package` to properly handle this optional search.

This is needed since `rapids_cpm_find` always records a version
value, but the project might not actually generate the correct
CMake infrastructure to support versioned `find_package` calls