Add support for a prefix in Cython module targets (#198)
A package composed of subpackages may contain multiple Cython modules with the same filename at different paths i.e. in different parts of the package directory tree. Since we use the filenames to construct the CMake library targets, the resulting library target names will conflict unless there is a way to customize them. This PR adds that functionality. Note that because of the way that scikit-build's Cython is designed, we need to ensure that the generated shared library is named  according to the original filename; the prefix should only be applied to the CMake target to avoid conflicts within CMake.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/rapids-cmake/pull/198