More robust solution of CMake policy 135 (#224)
By enabling CMake policy 135 we ensure that extracted files timestamp match that of the download time, instead of when the
archive is created. This makes sure that if the URL changes to an older version we still rebuild everything as the timestamp
stays newer.

Authors:
  - Robert Maynard (https://github.com/robertmaynard)

Approvers: None

URL: https://github.com/rapidsai/rapids-cmake/pull/224