Add ability to specify library directories for target rpaths (#295)
Python extension modules built using scikit-build and rapids-cmake are typically designed around a model where a C++ library is built and the resulting Cython extension modules must link to it. While it is in principle possible to use both scikit-build and rapids-cmake with only Cython, that is not the case for any of the major RAPIDS packages. When the C++ component is built via scikit-build and installed into the final package, it must also be discoverable by the extension modules. Currently this is accomplished by propagating a variable `RAPIDS_CYTHON_CREATED_MODULES` from `rapids_cython_create_modules` that calling functions can use to get the list of targets whose RPATHs need updating, and setting these RPATHs is the responsibility of the caller. However, this approach is error-prone and frequently results in incorrect RPATH specifications.

This PR automates the process by having the calling code provide a set of directories where compiled libraries may be placed. The RPATHs of all extension modules built via `rapids_cython_create_modules` are then automatically updated to include those library directories.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/rapids-cmake/pull/295