Move date to build string in `conda` recipe
This PR moves the date string from the version to the build string for conda recipes in this repository.

This is necessary to ensure that the conda packages resulting from PR builds can be installed in the same environment as nightly conda packages. This is useful for testing purposes.

Additionally, this PR adds the `PKG_HASH` value to our build string. This value is a hash computed from the dependency names and versions used to build our packages.

Typically the `PKG_HASH` value is included in packages when the build string is omitted, as seen on most `conda-forge` packages.

Including it in our build string will help ensure that new packages are published whenever our dependencies change (which is important in the case of shared library updates).

xref: https://github.com/rapidsai/rmm/pull/1195