rapids_cpm_nvbench properly specify usage of external fmt library
When we are inside a conda env the linker will be set to ld.bfd which will try to resolve all undefined symbols at time.

Since we could be using a shared library version of fmt we need it on the final link line of consumers. So patch nvbench to understand this requirement.