Fix google benchmark name and update version (#425)
The google benchmark package name was incorrect, making it impossible to find benchmark if it was previously installed. However, the currently pinned version of benchmark has a broken CMake config on conda-forge (it claims it is version 0.0.0) so it cannot be found even with the name fix. This is as good a time as any to go ahead and update to the latest version.

Resolves #424

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rapids-cmake/pull/425