Quote the list of patch files in case they have spaces in their paths (#463)
The rapids cpm scripts fail when populating a dependency with patches if the build directory contains spaces in its path name. The reason is because the paths are not properly quoted in the cmake patch script generated from `command_template.cmake.in`.

This PR properly quotes the list of patch file names.

Authors:
  - Eric Niebler (https://github.com/ericniebler)
  - Robert Maynard (https://github.com/robertmaynard)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rapids-cmake/pull/463