Express Python version in dependencies.yaml. (#470)
This PR adds `py_version` to the `dependencies.yaml` file so that `rapids-cmake` expresses this in the same way as every other RAPIDS repository. The primary reason this might matter is for ensuring that compatible Python versions are used for style checks.

Authors:
  - Bradley Dice (https://github.com/bdice)
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/rapids-cmake/pull/470