Support scikit-build-core (#433)
This PR adds a new cython-core module of rapids-cmake that can be used with scikit-build-core. `cython-core` is designed as a drop-in replacement for `cython`, allowing consumers to simply change the `include(cython)` to `include(cython-core)` and have things generally work as expected when switching Python build backends.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Robert Maynard (https://github.com/robertmaynard)

URL: https://github.com/rapidsai/rapids-cmake/pull/433