Officially support env var expansion in version.json
Expand rapids-cmake to support env var expansion in the
git_url, git_tag, version, and binary entries.

This allows rapids-cmake to support users that need to override
versions.json with url info that needs secure checkout data provided
by the running container. This is commonly due to cloning from
a private repo that has user and password setup in the running
docker container.