Revert CMP0169 before CPM usage to support CMake 3.30
When a project has specified CMake 3.30 as the minimum required
value it forces `CMP0169` to be a hard error.

So when need to mark it as OLD for now while we bump the minimum
required CMake version in each RAPIDS project, before removing
this when we update our CPM dependency version.