Update gtest version (#795)
The version of gtest that we are using (1.13) is no longer compatible with the version of gtest that is default in conda-forge (1.15.2). The issue is in "private APIs" changing: gtest is intended to be used via its macros, so directly calling some of the functions whose APIs have changed is not supported, but the problem is that the macros produce code using those functions. The second problem is that gtest and gmock mark their includes as isystem when they are built from source, which means that when building inside any environment where gtest is installed to system locations, if that version of gtest has incompatible versions of these APIs we wind up using the headers from the system installation (e.g. conda's) but then we link to our local build, resulting in linker errors. That is the issue that is currently arising. Due to the number of warnings that would surface if we tried to change the `-isystem` to `-I` for gtest, our best option is to just update gtest rather than muck with any of that. Since all of RAPIDS builds a static copy of gtest and avoids bundling it now, it should be safe to upgrade the version now.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rapids-cmake/pull/795