Update to NVTX 3.2.0. (#844)
This updates to NVTX 3.2.0. This was already tested downstream with RMM and cuDF before the final release was made. We were previously using a pre-release of 3.2.0 so no significant changes are expected. Thus, I think this should be safe to merge without downstream testing.

Authors:
  - Bradley Dice (https://github.com/bdice)
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

Approvers:
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

URL: https://github.com/rapidsai/rapids-cmake/pull/844