rapids-cmake: Add support for a version suffix to mean using main (#864)
To support the new RAPIDS wide branching model rapids-cmake has to support switching between the `main` branch for `rapids_cpm_rmm` and a release branch like `branch-25.06`.

This is implemented by reading the rapids-cmake `VERSION` file and when we detect a value that ends in `a` we know we are being used from the `main` branch (a development version) and not from a release branch.

Fixes #863

Authors:
  - Robert Maynard (https://github.com/robertmaynard)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rapids-cmake/pull/864